Universally cleaned up numeric read/writes with datafiles Added basic concession lookup for $ALLUSERSPROFILE/Application Data/ Fixed personal home lookup on windoze to use non-posix $HOMEDRIVE$HOMEPATH Added concession to try ~/sword (after ~/.sword)
git-svn-id: https://crosswire.org/svn/sword/trunk@2240 bcd7d363-81e1-0310-97ec-a550e20fc99c

+++ include/rawstr.h
@@ -26,6 +26,7 @@
 #define RAWSTR_H
 
 #include <defs.h>
+#include <sysdata.h>
 
 SWORD_NAMESPACE_START
 
@@ -48,8 +49,8 @@ public:
 	virtual ~RawStr();
 	void getIDXBuf(long ioffset, char **buf);
 	void getIDXBufDat(long ioffset, char **buf);
-	signed char findOffset(const char *key, long *start, unsigned short *size, long away = 0, long *idxoff = 0);
-	void readText(long start, unsigned short *size, char **idxbuf, SWBuf &buf);
+	signed char findOffset(const char *key, __u32 *start, __u16 *size, long away = 0, __u32 *idxoff = 0);
+	void readText(__u32 start, __u16 *size, char **idxbuf, SWBuf &buf);
 	static signed char createModule(const char *path);
 };
